Moved components to their correct locations
